extended support for idle_a, idle_b, idle_c states
Added support for idle_a, idle_b, idle_c in addition to idle, as options in the conf file.
I.e. Seagate Exos drives switch to idle_a almost instantaneous with almost no performance penalty for switching back to active. In idle_b state the head is parked, in idle_c the rotation speed is reduced, switching back to active incurs a ~2 seconds penalty. It is therefore indispensable to be able to discern those states for disk check (i.e. check disk in idle_a state, while skipping in idle_c state).
The modification maintains backwards compatibility as the new states are encoded as cfg.powermode= 4 / 5 / 6 and those numbers satisfy the ">=3" if-condition that is used in all lines that check for idle-state. This encoding maintains the "higher number = closer to full active" sheme.