smartd: Add device failure check in smartd
This patch is used to enhance device failure detection and alarm in
smartd. This alarm is reported when the system detects that an indicator
is abnormal compared with the configured value.
It mainly includes the following:
(1) Wear proportion monitoring. This value is read directly from the
device and compared with the set threshold. When the wear ratio is
too large, an email alarm is reported.
(2) Monitor the remaining reserved space. This value is also mainly read
directly from the device and compared with the set threshold. When the
wear ratio is too large, an email alarm is reported.
(3) Remaining life time monitoring. The value is calculated using the
following formula:
Workload Remained Time = (POH x Remained wear ratio)/Weared wear ratio.
When the remaining time is less than the preset threshold, an alarm is generated.
(4) Wear speed monitoring. The value is calculated using the following formula:
Workload Wear Speed = Write data volume/(Capacity of a single disk x POH).
When the wear speed is greater than the preset threshold, an alarm is generated.

Signed-off-by: YunYi Yang <yangyunyi2@huawei.com>