nvmeprint.cpp: Fix self-test for devices with multiple namespaces
According to the NVMe specs, the 06h log page has only Controller and
NVM Subsystem operation scopes, but not a namespace. Attempting to read
this log page with a specific device NSID results in an `Invalid Field
in Command` error.
This issue was previously resolved for single-namespace devices by using
the broadcast NSID. However, the problem persists for devices with
multiple namespaces, where the log page is still being accessed with a
device-specific NSID. This fix ensures that the broadcast NSID is always
used when reading the 06h log page.